+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - adds `getScreenPosition` method in `GameComponent`.
 - adds `enableDiagonal` param in `simpleAttackRange`.
 - adds `visibleComponentsByType` and `componentsByType` in `BonfireGame`.
+- adds `onTapDown` in `BonfireTiledWidget` and `BonfireWidget`.
 
 # 1.5.11
 - Improvements performance.